Structure the game a little better.
